@@ -1,3 +1,3 @@
-@using CoreTest
+﻿@using CoreTest
 @using CoreTest.Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
